starter-code-simple/requirements.txt
Flask==2.3.2
requests==2.31.0
pytest==7.4.2
flake8==6.1.0
black==24.3.0
bandit==1.8.1
python-dotenv==1.0.0
bcrypt=4.0.1
